+++ cmd/add/location_test.go
@@ -7,18 +7,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TestGetLocationCmd_Returns tests that GetLocationCmd returns a non-nil command.
-func TestGetLocationCmd_Returns(t *testing.T) {
-	cmd := GetLocationCmd()
-	require.NotNil(t, cmd)
-	assert.Contains(t, cmd.Use, "location")
-	assert.NotEmpty(t, cmd.Short)
-	assert.NotEmpty(t, cmd.Long)
-}
-
-// TestGetLocationCmd_Singleton tests that GetLocationCmd returns the same instance.
-func TestGetLocationCmd_Singleton(t *testing.T) {
+func TestGetLocationCmd(t *testing.T) {
 	cmd1 := GetLocationCmd()
+	require.NotNil(t, cmd1)
+
 	cmd2 := GetLocationCmd()
+
 	assert.Same(t, cmd1, cmd2)
 }
