+++ cmd/config/edit_test.go
@@ -3,68 +3,15 @@ package config
 import (
 	"testing"
 
-	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func TestConfigEdit_LocalFlag(t *testing.T) {
-	defer ResetForTesting()
+func TestGetEditCmd_Singleton(t *testing.T) {
+	cmd1 := GetEditCmd()
+	require.NotNil(t, cmd1)
 
-	cmd := GetEditCmd()
-	local := cmd.Flags().Lookup("local")
+	cmd2 := GetEditCmd()
 
-	assert.NotNil(t, local)
-}
-
-func TestConfigEdit_GlobalFlag(t *testing.T) {
-	defer ResetForTesting()
-
-	cmd := GetEditCmd()
-	global := cmd.Flags().Lookup("global")
-
-	assert.NotNil(t, global)
-}
-
-func TestConfigEdit_HasRunE(t *testing.T) {
-	defer ResetForTesting()
-
-	cmd := GetEditCmd()
-
-	assert.NotNil(t, cmd.RunE)
-}
-
-func TestConfigEdit_ReturnsWithoutEditor(t *testing.T) {
-	defer ResetForTesting()
-
-	// Skip this test - it tries to open an editor which hangs in CI/test environment
-	t.Skip("Skipping editor test - requires EDITOR environment variable and interactive shell")
-}
-
-func TestConfigEdit_GlobalPathUsed(t *testing.T) {
-	defer ResetForTesting()
-
-	memFS := afero.NewMemMapFs()
-	SetFilesystem(memFS)
-	defer SetFilesystem(afero.NewOsFs())
-
-	cmd := GetEditCmd()
-
-	// Verify command structure
-	assert.NotNil(t, cmd)
-	assert.Equal(t, "edit", cmd.Use)
-}
-
-func TestConfigEdit_LocalPathUsed(t *testing.T) {
-	defer ResetForTesting()
-
-	memFS := afero.NewMemMapFs()
-	SetFilesystem(memFS)
-	defer SetFilesystem(afero.NewOsFs())
-
-	cmd := GetEditCmd()
-	cmd.SetArgs([]string{"--local"})
-
-	// Verify command structure
-	assert.NotNil(t, cmd)
-	assert.Equal(t, "edit", cmd.Use)
+	assert.Same(t, cmd1, cmd2)
 }
