+++ cmd/config/set_test.go
@@ -7,42 +7,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestConfigSet_ReturnsCommand(t *testing.T) {
-	defer ResetForTesting()
+func TestGetSetCmd_Singleton(t *testing.T) {
+	cmd1 := GetSetCmd()
+	require.NotNil(t, cmd1)
 
-	cmd := GetSetCmd()
+	cmd2 := GetSetCmd()
 
-	require.NotNil(t, cmd)
-	assert.NotEmpty(t, cmd.Use)
-}
-
-func TestConfigSet_LocalFlag(t *testing.T) {
-	defer ResetForTesting()
-
-	cmd := GetSetCmd()
-	local := cmd.Flags().Lookup("local")
-
-	assert.NotNil(t, local)
-}
-
-func TestConfigSet_NoGlobalFlag(t *testing.T) {
-	defer ResetForTesting()
-
-	cmd := GetSetCmd()
-
-	// Set command only has --local flag, not --global
-	// By default it modifies global config
-	local := cmd.Flags().Lookup("local")
-	assert.NotNil(t, local)
-
-	global := cmd.Flags().Lookup("global")
-	assert.Nil(t, global)
-}
-
-func TestConfigSet_HasRunE(t *testing.T) {
-	defer ResetForTesting()
-
-	cmd := GetSetCmd()
-
-	assert.NotNil(t, cmd.RunE)
+	assert.Same(t, cmd1, cmd2)
 }
