+++ cmd/initialize/initialize_test.go
@@ -1,4 +1,4 @@
-package config
+package initialize
 
 import (
 	"testing"
@@ -7,11 +7,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestConfigCheck(t *testing.T) {
-	cmd1 := GetConfigCmd()
+func TestGetInitializeCmd_Singleton(t *testing.T) {
+	cmd1 := GetInitializeCmd()
 	require.NotNil(t, cmd1)
 
-	cmd2 := GetConfigCmd()
+	cmd2 := GetInitializeCmd()
 
 	assert.Same(t, cmd1, cmd2)
 }
