+++ cmd/move/item_test.go
@@ -388,7 +388,7 @@ func TestResult_JSONMarshal(t *testing.T) {
 
 // Test: Command structure and flag parsing.
 func TestGetMoveCmd_Structure(t *testing.T) {
-	cmd := GetMoveCmd()
+	cmd := NewDefaultMoveCmd()
 
 	assert.NotNil(t, cmd)
 	assert.Equal(t, "move", cmd.Name())
