+++ flake.nix
@@ -107,7 +107,7 @@
             mise trust --all
           '';
 
-          CGO_ENABLED = "0";
+          CGO_ENABLED = "1";
         };
       }
     )
